category doughnut now reflects current records table info
